fix: failing to parse an existing google-services.json file throws exception
An google-services.json file may exist but may not always be able to be parsed for a number of reasons such as an empty file, or corrupt content.

Rather than throwing an exception and aborting the command, if an exception occurs while trying to extract to parse the existing google-services.json file, we now prompt the user to overwrite the file.