fix: configure command not updating build.gradle files
The issue for this bug was that the _androidBuildGradleRegex regex expression was failing when it tried to match against the document, hence it would have failed even in the replaceFirstMapped method later on in the code.

Changed the regex to capture only the `classpath 'com.android.tools.build:gradle:7.1.2'` with the indentation included, and also modified the replaceFirstMapped accordingly.

closes issue #126